% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pull_and_fill.R
\name{data_pull_and_fill}
\alias{data_pull_and_fill}
\title{Fills in Sequences of input file}
\usage{
data_pull_and_fill(input_file, ref_protein, ncbi_bin, num = 7)
}
\arguments{
\item{input_file}{Input data file}

\item{ref_protein}{Path to Fasta refrence file}

\item{ncbi_bin}{Path to ncbi-blast-2.14.0+/bin}

\item{num}{Number of amino acids to extract on either side of the modified amino acid (default = 7)}
}
\value{
Final Data Sheet w/ Extracted Sequences
}
\description{
Loops through reformatted input file and pulls 15 character ranges from blast.
}
