% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatter.R
\name{file_reformater}
\alias{file_reformater}
\title{Format File Input}
\usage{
file_reformater(input_path, num = 7)
}
\arguments{
\item{input_path}{Path to input file}

\item{num}{Number of amino acids to extract on either side of the modified amino acid (default = 7)}
}
\value{
Reformatted Input File
}
\description{
Parse and reformat input file to have only 1 row corresponding to 1 accession.
}
