% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_blast_plus_binaries.R
\name{has_blast_plus_binaries}
\alias{has_blast_plus_binaries}
\title{Check BLAST+ installation}
\usage{
has_blast_plus_binaries()
}
\value{
TRUE/FALSE with attributes ncbi_bin, checked_binaries
}
\description{
Checks required BLAST+ binaries are in the PATH, or found with given path. This
function returns a TRUE/FALSE for quick logic.
}