% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastdbcmd.R
\name{run_blastdbcmd}
\alias{run_blastdbcmd}
\title{Run Blast Database Command}
\usage{
run_blastdbcmd(query_row, db, ncbi_bin)
}
\arguments{
\item{query_row}{Row of interest}

\item{db}{Path to Fasta refrence file}

\item{ncbi_bin}{Path to ncbi-blast-2.14.0+/bin}
}
\description{
Pulls protein sequence from database for given row
}
