% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_motif_extraction.R
\name{run_motif_extraction}
\alias{run_motif_extraction}
\title{Motif Extraction}
\usage{
run_motif_extraction(input_path, output_path, ncbi_bin, protein_ref, num = 7)
}
\arguments{
\item{input_path}{Path to input file}

\item{output_path}{Path to output file}

\item{ncbi_bin}{Path to ncbi-blast-2.14.0+/bin}

\item{protein_ref}{Path to Fasta refrence file}

\item{num}{Number of amino acids to extract on either side of the modified amino acid (default = 7)}
}
\value{
Input file with addition of protein sequences
}
\description{
Reformat data file, build Blast Database, pull sequences and add output.
}
