% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpFileGenerator.R
\docType{package}
\name{UVMProteomics}
\alias{UVMProteomics}
\title{UVMProteomics}
\description{
This package is designed for the UVM Proteomics Core to finds the
7 residues before and after given accessions and phosphosites of interest.
}
\author{
Emily Guswa \email{Emily.Guswa@uvm.edu}, James (Jax) Lubkowitz \email{jalubk20@stlawu.edu}
}
