% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_blastdb.R
\name{build_blastdb}
\alias{build_blastdb}
\title{Build Blast Database}
\usage{
build_blastdb(ref_protein, ncbi_bin)
}
\arguments{
\item{ref_protein}{Path to Fasta refrence file}

\item{x}{Protein Sequence for the Reference Database}
}
\description{
Generated blast database from NCBI
}
